refactor: Launch base class refactor
