+++ packages/contentstack-launch/src/commands/launch/deployments.ts
@@ -4,12 +4,11 @@ import find from 'lodash/find';
 import isEmpty from 'lodash/isEmpty';
 import { FlagInput, Flags, cliux as ux } from '@contentstack/cli-utilities';
 
-import { BaseCommand } from './base-command';
+import { BaseCommand } from '../../base-command';
 import { environmentsQuery } from '../../graphql';
 import { Logger, selectOrg, selectProject } from '../../util';
 
 export default class Deployments extends BaseCommand<typeof Deployments> {
-  static hidden = false;
   static description = 'Show list of deployments for an environment';
 
   static examples = [
